+++ NEWS.md
@@ -15,12 +15,15 @@ the CSV. (#394)
 troubleshooting, debugging and development. See end of *How does CmdStanR work?*
 vignette for details. (#392)
 
-* New `loo()` method for CmdStanMCMC objects. Requires computing pointwise
+* New `$loo()` method for CmdStanMCMC objects. Requires computing pointwise
 log-likelihood in Stan program. (#366)
 
 * The `fitted_params` argument to the `$generate_quantities()` method now also
 accepts CmdStanVB, `posterior::draws_array`, and `posterior::draws_matrix`
 objects. (#390)
+
+* The `$optimize()` method now supports all of CmdStan's tolerance-related
+arguments for (L)BFGS. (#398)
 
 # cmdstanr 0.2.2